// Module Manager : convert module import to the new archi
